All Lure No Lore
I was thinking about possible story elements, but ended up mostly playing with lures.  Couple other tweeks:
 - input strings are made lowercase to remove casesensitivity.   Similarly, fish are checked titlecase.
 - Fish heal 1 + the fish's name length mod three-- some variation in hp recovery
 - made some tweeks to health/damage checks to insure player's HP stays between zero and max hp
 - Bread, lure, and worm can be used with the fishing rod to make the bread rod, lured rod, and worm rod respectively.  These all have different stats for catching fish.  Some new catch conditions are "trade" (lure is lost, but fish is caught), "snag" (lure is lost).  Maybe lures can change the speed of the cast pole characters?

Oh, there's a chance to find bait (bread/worms) every time the player travels to a new location now too.

Let me know what you think!  I'm not particularly happy with all the ifs in equiptLure()-- we'll need a better way for the equipt prompt to only offer items in the players inventory, especaily if we add more lures!

So happy to be working on this!  What can we do with that bow and arrow?
Cheers,
 - Ben